pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dummy-a2a"
version = "0.1.0"
description = "A dummy a2a agent for testing."
readme = "README.md"
requires-python = ">=3.12"
license = "Apache-2.0"
license-files = ["LICENSE"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
    "Typing :: Typed",
]
dependencies = [
    "a2a-sdk[http-server,sqlite]==1.0.0a0",
    "uvicorn>=0.34",
    "httpx>=0.28",
]
authors = [
    { name = "agsuy", email = "37564412+agsuy@users.noreply.github.com" },
]

[project.urls]
Repository = "https://github.com/agsuy/dummy-a2a"
Issues = "https://github.com/agsuy/dummy-a2a/issues"

[project.scripts]
dummy-a2a = "dummy_a2a.cli:main"

# [project.optional-dependencies]
# example = ["some-package>=1.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
dummy_a2a = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "."]

[tool.coverage.run]
source = ["dummy_a2a"]

[tool.coverage.report]
fail_under = 90
show_missing = true

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "pytest-cov>=6.0",
    "python-semantic-release>=10.5",
    "ruff>=0.15.6",
    "pytest-asyncio>=0.26",
    "cryptography>=44.0",
]

[tool.semantic_release]
commit_author = "agsuy <37564412+agsuy@users.noreply.github.com>"
commit_message = "chore(release): {version}\n\n[skip ci]"
commit_parser = "conventional"
tag_format = "v{version}"
allow_zero_version = true
major_on_zero = false
version_toml = ["pyproject.toml:project.version"]
version_variables = ["src/dummy_a2a/__init__.py:__version__"]
build_command = """
uv lock --upgrade-package "$PACKAGE_NAME"
git add uv.lock
uv build
"""

[tool.semantic_release.branches.main]
match = "(main)"
prerelease = false

[tool.semantic_release.changelog.default_templates]
changelog_file = "CHANGELOG.md"
mask_initial_release = true

[tool.semantic_release.remote]
type = "github"

[tool.semantic_release.remote.token]
env = "GH_TOKEN"

[tool.semantic_release.publish]
upload_to_vcs_release = true

[tool.ruff]
target-version = "py312"
line-length = 100
indent-width = 4
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I"]
